Create Fundamentals.py
This is a decent warm up for beginners. This is not intended to be a full course with a solid curriculum, rather a good warm up to do each day to really drive home the fundamentals and expose anything you may have not learned yet.

If you are a more experienced user of either Github or Python, I would greatly appreciate any feedback. 

Thanks for reading and have a beautiful day.